% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivo_table_gt.R
\encoding{UTF-8}
\name{ivo_gt_theme}
\alias{ivo_gt_theme}
\title{A nice GT theme}
\usage{
ivo_gt_theme(table, color = "darkgreen", font_name = "Arial")
}
\arguments{
\item{table}{A \code{gt} object.}

\item{color}{A named color or a color HEX code, used for the lines in the table. Defaults to "darkgreen".}

\item{font_name}{The name of the font to be used in the table. Defaults to "Arial".}
}
\value{
A styled \code{gt} table.
}
\description{
Applies a predefined theme to a GT table, including font, colors, and alignment.
}
\author{
Stefan Furne
}
