% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivo_table_gt.R
\encoding{UTF-8}
\name{ivo_table_gt}
\alias{ivo_table_gt}
\title{Create pretty frequency/contingency GT tables}
\usage{
ivo_table_gt(
  df,
  title = NULL,
  subtitle = NULL,
  extra_header = TRUE,
  source_note = NULL,
  mask = FALSE,
  missing_string = "(Missing)",
  totals = NULL
)
}
\arguments{
\item{df}{A data frame with 1-3 columns}

\item{title}{An optional string containing a table title.}

\item{subtitle}{An optional string containing a table subtitle. Only usable together with title.}

\item{extra_header}{Should the variable name be displayed? Defaults to TRUE.}

\item{source_note}{An optional string for a table source note.}

\item{mask}{WIP}

\item{missing_string}{A string used to indicate missing values. Defaults to "(Missing)".}

\item{totals}{WIP}
}
\value{
A stylized \code{GT} table.
}
\description{
\code{ivo_table_gt()} lets you easily create a GT table using pretty fonts and colors.
}
\details{
The functions \code{ivo_table_gt()} takes a \code{data.frame} with 1-3 columns. The order of the columns in the \code{data.frame} will determine where they will be displayed in the table. The first column will always be displayed at the top of the table. If there are more than one column the following 2-3 columns will be displayed to the left in order. To change how the columns are displayed in the table; change the place of the columns in the \code{data.frame} using \code{dplyr::select()}.
}
\author{
Stefan Furne
}
